dev-requirements.txt
-r headless_browser/requirements.txt
ruff==0.6.7
pre-commit==3.8.0
pytest==8.3.3
pytest-asyncio==0.23.6
python-dotenv>=1.0.0
